Extract format logic to new Github::Nippou::Format class
