+++ lib/github/nippou.rb
@@ -2,5 +2,6 @@ require 'github/nippou/concerns/sawyer_resource_github'
 require 'github/nippou/concerns/string_markdown'
 
 require 'github/nippou/commands'
+require 'github/nippou/format'
 require 'github/nippou/user_events'
 require 'github/nippou/version'
